Added site prep, copyFile()
